fix(web): changing use of auth0 internal to web
createAuth0Client was calling getTokenSilently(), which in browsers without ITP turned on was
causing users to appear logged in even if they were not actually logged in the game itself. The
interface for build was async and due to the timing a sleep was put in to allow an await to be
utilized. This will be fixed in the next minor version post launch.