fix(web): updated web to local storage
Ultimately determined that certain browsers like Safari and likely Brave just will not work with the
SPAs unless we use the local storage option from Auth0. It should be a low risk since we have
refresh tokens turned on and Auth0 can automatically detect when a token is being reused.