feat(web): updated web sdk to next version
This is a breaking change updating to the newest auth where a game developer is no longer required
to have a separate web client id.

BREAKING CHANGE: This is a breaking change updating to the newest auth where a game developer is no
longer required to have a separate web client id.